Added string “bigamount” to replace “amount” int64
