+++ src/opentxs-proto/SendPayment.proto
@@ -20,4 +20,5 @@ message SendPayment {
     optional string destinationaccount = 5;	// destination account id
     optional string memo = 6;
     optional uint64 amount = 7;
+    optional string amount_mp = 8;
 }
